feat: updates to make code more readable
