+++ src/squishHelpers/id.js
@@ -1,0 +1,16 @@
+const ID_SUBTYPE = 43;
+
+const squishId = {
+	type: ID_SUBTYPE,
+	squish: (i) => {
+		return [i];
+	},
+	unsquish: (arr) => {
+		return arr[0];
+	}
+}
+
+module.exports = {
+    ID_SUBTYPE,
+    squishId
+};
