+++ src/squishHelpers/subType.js
@@ -1,0 +1,16 @@
+const SUBTYPE_TYPE = 55;
+
+const squishSubType = {
+	type: SUBTYPE_TYPE,
+	squish: (a) => {
+		return [a];
+	},
+	unsquish: (s) => {
+		return s[0];
+	}
+}
+
+module.exports = {
+    SUBTYPE_TYPE,
+    squishSubType
+};
